Documents lab setup steps (#4)
TL;DR
-----

Adds documentation of setup procedures

Details
-------

Explains step-by-step how to setup the Lab. The docs lean on what came
before in the TKG Lab and in @doddatpivotal's end-to-end lab adaptation
of that code. 

Like its predecessors, these docs show a step-by-step guide to getting 
the lab configured to run a demonstration. Most of the work is automated
with a set of scripts in the `bin` directory, some of which I added as
I documented the process. Even the scripts I had written before got a
few updates to make them more robust.

These docs do not explain what to do once you have everything up and
running. Those docs are coming next.