Automates TMC policy configuration (#5)
TL;DR
-----

Adds a script to setup TMC policies for the lab

Details
-------

Introduces a script `prepare-tmc-policy.sh` that sets up as much
of the TMC policy environment for the lab as possible. It doesn't
do all of the setup since it's not apparent how to assign a 
custom policy from the CLI. If its possible, it's not readily
discoverable. It may also not yet be implemented as the CLI has
been known to lag the GUI.

Also updates the documentation to explain how to set up TMC policy 
both with the script and the manual steps that are currently
required.